implement stock list, single stock api
