+++ stocks/serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from reports.models import Stock
+
+
+class StockSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Stock
+        fields = ["id", "name", "code"]
