+++ stocks/urls.py
@@ -1,7 +1,14 @@
 from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from reports.views import StockReportsView, StockViewSet, PointViewSet
+from rest_framework import routers
+from reports.views import StockReportsView
+from stocks.views import StockViewSet
+
+app_name = "stocks"
+
+router = routers.DefaultRouter()
+router.register("", StockViewSet)
 
 urlpatterns = [
+    path("", include(router.urls)),
     path("<int:pk>/reports/", StockReportsView.as_view(), name="stock-reports"),
 ]
